net: framework: Move proto from framework/net
