--- src/framework/net/proto/Mybuild
@@ -1,7 +1,0 @@
-package embox.framework.net
-
-module proto {
-	source "core.c"
-
-	depends embox.util.Array
-}
