+++ src/net/l3/Mybuild
@@ -16,7 +16,7 @@ module arp {
 module icmpv4 {
 	source "icmpv4.c"
 
-	depends embox.framework.net.proto
+	depends proto
 	depends embox.compat.libc.assert
 	depends embox.net.lib.icmpv4
 	depends embox.net.lib.ipv4
@@ -26,7 +26,7 @@ module icmpv4 {
 module icmpv6 {
 	source "icmpv6.c"
 
-	depends embox.framework.net.proto
+	depends embox.net.proto
 	depends embox.compat.libc.assert
 	depends embox.compat.libc.str
 	depends embox.net.core
@@ -66,4 +66,10 @@ module route {
 	depends core /* for inetdev.c */
 	depends embox.mem.pool
 	depends embox.util.DList
+}
+
+module proto {
+	source "proto.c"
+
+	depends embox.util.Array
 }
